refactor sql classes as SqlProcessors
