--- airbyte/_file_writers/__init__.py
@@ -1,16 +1,0 @@
-from __future__ import annotations
-
-from .base import FileWriterBase, FileWriterBatchHandle, FileWriterConfigBase
-from .jsonl import JsonlWriter, JsonlWriterConfig
-from .parquet import ParquetWriter, ParquetWriterConfig
-
-
-__all__ = [
-    "FileWriterBatchHandle",
-    "FileWriterBase",
-    "FileWriterConfigBase",
-    "JsonlWriter",
-    "JsonlWriterConfig",
-    "ParquetWriter",
-    "ParquetWriterConfig",
-]
