+++ airbyte/_processors/file/base.py
@@ -1,6 +1,5 @@
 # Copyright (c) 2023 Airbyte, Inc., all rights reserved.
-
-"""Define abstract base class for File Writers, which write and read from file storage."""
+"""Abstract base class for File Writers, which write and read from file storage."""
 
 from __future__ import annotations
 
@@ -11,8 +10,8 @@ from typing import TYPE_CHECKING, cast, final
 
 from overrides import overrides
 
-from airbyte._processors import BatchHandle, RecordProcessor
-from airbyte.config import CacheConfigBase
+from airbyte._processors.base import BatchHandle, RecordProcessor
+from airbyte.caches.base import CacheBase
 
 
 if TYPE_CHECKING:
@@ -34,7 +33,7 @@ class FileWriterBatchHandle(BatchHandle):
     files: list[Path] = field(default_factory=list)
 
 
-class FileWriterConfigBase(CacheConfigBase):
+class FileWriterConfigBase(CacheBase):
     """Configuration for the Snowflake cache."""
 
     cache_dir: Path = Path("./.cache/files/")
@@ -45,8 +44,6 @@ class FileWriterConfigBase(CacheConfigBase):
 
 class FileWriterBase(RecordProcessor, abc.ABC):
     """A generic base implementation for a file-based cache."""
-
-    config: FileWriterConfigBase
 
     @abc.abstractmethod
     @overrides
@@ -90,7 +87,7 @@ class FileWriterBase(RecordProcessor, abc.ABC):
 
         This method is a no-op if the `cleanup` config option is set to False.
         """
-        if self.config.cleanup:
+        if self.cache.cleanup:
             batch_handle = cast(FileWriterBatchHandle, batch_handle)
             _ = stream_name, batch_id
             for file_path in batch_handle.files:
