+++ airbyte/_processors/file/parquet.py
@@ -1,6 +1,5 @@
-# Copyright (c) 2023 Airbyte, Inc., all rights reserved.
-
-"""A Parquet cache implementation.
+# Copyright (c) 2023 Airbyte, Inc., all rights reserved
+"""A Parquet file writer implementation.
 
 NOTE: Parquet is a strongly typed columnar storage format, which has known issues when applied to
 variable schemas, schemas with indeterminate types, and schemas that have empty data nodes.
@@ -18,7 +17,7 @@ from overrides import overrides
 from pyarrow import parquet
 
 from airbyte import exceptions as exc
-from airbyte._file_writers.base import (
+from airbyte._processors.file.base import (
     FileWriterBase,
     FileWriterBatchHandle,
     FileWriterConfigBase,
@@ -42,7 +41,7 @@ class ParquetWriter(FileWriterBase):
     ) -> Path:
         """Return a new cache file path for the given stream."""
         batch_id = batch_id or str(ulid.ULID())
-        config: ParquetWriterConfig = cast(ParquetWriterConfig, self.config)
+        config: ParquetWriterConfig = cast(ParquetWriterConfig, self.cache)
         target_dir = Path(config.cache_dir)
         target_dir.mkdir(parents=True, exist_ok=True)
         return target_dir / f"{stream_name}_{batch_id}.parquet"
