+++ airbyte/_processors/sql/__init__.py
@@ -1,0 +1,2 @@
+# Copyright (c) 2023 Airbyte, Inc., all rights reserved.
+"""SQL processors."""
