+++ airbyte/caches/__init__.py
@@ -1,7 +1,8 @@
+# Copyright (c) 2023 Airbyte, Inc., all rights reserved.
 """Base module for all caches."""
 from __future__ import annotations
 
-from airbyte.caches.base import SQLCacheBase
+from airbyte.caches.base import CacheBase
 from airbyte.caches.duckdb import DuckDBCache
 from airbyte.caches.postgres import PostgresCache
 from airbyte.caches.snowflake import SnowflakeCache
@@ -11,6 +12,6 @@ from airbyte.caches.snowflake import SnowflakeCache
 __all__ = [
     "DuckDBCache",
     "PostgresCache",
-    "SQLCacheBase",
+    "CacheBase",
     "SnowflakeCache",
 ]
