+++ airbyte/results.py
@@ -12,12 +12,15 @@ if TYPE_CHECKING:
 
     from sqlalchemy.engine import Engine
 
-    from airbyte.caches import SQLCacheBase
+    from airbyte.caches import CacheBase
 
 
 class ReadResult(Mapping[str, CachedDataset]):
     def __init__(
-        self, processed_records: int, cache: SQLCacheBase, processed_streams: list[str]
+        self,
+        processed_records: int,
+        cache: CacheBase,
+        processed_streams: list[str],
     ) -> None:
         self.processed_records = processed_records
         self._cache = cache
@@ -52,5 +55,5 @@ class ReadResult(Mapping[str, CachedDataset]):
         }
 
     @property
-    def cache(self) -> SQLCacheBase:
+    def cache(self) -> CacheBase:
         return self._cache
