+++ tests/integration_tests/test_source_test_fixture.py
@@ -9,7 +9,7 @@ from typing import Any
 from unittest.mock import Mock, call, patch
 import tempfile
 from pathlib import Path
-from airbyte.caches.base import SQLCacheInstanceBase
+from airbyte._processors.sql.base import SqlProcessorBase
 
 from sqlalchemy import column, text
 
@@ -202,7 +202,7 @@ def test_file_write_and_cleanup() -> None:
         assert len(list(Path(temp_dir_2).glob("*.jsonl.gz"))) == 2, "Expected files to exist"
 
 
-def assert_cache_data(expected_test_stream_data: dict[str, list[dict[str, str | int]]], cache: SQLCacheInstanceBase, streams: list[str] = None):
+def assert_cache_data(expected_test_stream_data: dict[str, list[dict[str, str | int]]], cache: SqlProcessorBase, streams: list[str] = None):
     for stream_name in streams or expected_test_stream_data.keys():
         if len(cache[stream_name]) > 0:
             pd.testing.assert_frame_equal(
@@ -292,13 +292,13 @@ def test_read_isolated_by_prefix(expected_test_stream_data: dict[str, list[dict[
     db_path = Path(f"./.cache/{cache_name}.duckdb")
     source = ab.get_source("source-test", config={"apiKey": "test"})
     source.select_all_streams()
-    cache = ab.DuckDBCacheInstance(config=ab.DuckDBCache(db_path=db_path, table_prefix="prefix_"))
+    cache = ab.DuckDBSqlProcessor(config=ab.DuckDBCache(db_path=db_path, table_prefix="prefix_"))
 
     source.read(cache)
 
-    same_prefix_cache = ab.DuckDBCacheInstance(config=ab.DuckDBCache(db_path=db_path, table_prefix="prefix_"))
-    different_prefix_cache = ab.DuckDBCacheInstance(config=ab.DuckDBCache(db_path=db_path, table_prefix="different_prefix_"))
-    no_prefix_cache = ab.DuckDBCacheInstance(config=ab.DuckDBCache(db_path=db_path, table_prefix=None))
+    same_prefix_cache = ab.DuckDBSqlProcessor(config=ab.DuckDBCache(db_path=db_path, table_prefix="prefix_"))
+    different_prefix_cache = ab.DuckDBSqlProcessor(config=ab.DuckDBCache(db_path=db_path, table_prefix="different_prefix_"))
+    no_prefix_cache = ab.DuckDBSqlProcessor(config=ab.DuckDBCache(db_path=db_path, table_prefix=None))
 
     # validate that the cache with the same prefix has the data as expected, while the other two are empty
     assert_cache_data(expected_test_stream_data, same_prefix_cache)
@@ -310,9 +310,9 @@ def test_read_isolated_by_prefix(expected_test_stream_data: dict[str, list[dict[
     source.read(different_prefix_cache)
     source.read(no_prefix_cache)
 
-    second_same_prefix_cache = ab.DuckDBCacheInstance(config=ab.DuckDBCache(db_path=db_path, table_prefix="prefix_"))
-    second_different_prefix_cache = ab.DuckDBCacheInstance(config=ab.DuckDBCache(db_path=db_path, table_prefix="different_prefix_"))
-    second_no_prefix_cache = ab.DuckDBCacheInstance(config=ab.DuckDBCache(db_path=db_path, table_prefix=None))
+    second_same_prefix_cache = ab.DuckDBSqlProcessor(config=ab.DuckDBCache(db_path=db_path, table_prefix="prefix_"))
+    second_different_prefix_cache = ab.DuckDBSqlProcessor(config=ab.DuckDBCache(db_path=db_path, table_prefix="different_prefix_"))
+    second_no_prefix_cache = ab.DuckDBSqlProcessor(config=ab.DuckDBCache(db_path=db_path, table_prefix=None))
 
     # validate that the first cache still has full data, while the other two have partial data
     assert_cache_data(expected_test_stream_data, second_same_prefix_cache)
@@ -423,7 +423,7 @@ def test_cached_dataset(
     not_a_stream_name = "not_a_stream"
 
     # Check that the stream appears in mapping-like attributes
-    assert stream_name in result.cache._streams_with_data
+    assert stream_name in result.cache._expected_streams
     assert stream_name in result
     assert stream_name in result.cache
     assert stream_name in result.cache.streams
