+++ tests/unit_tests/test_caches.py
@@ -4,10 +4,10 @@ from pathlib import Path
 
 import pytest
 
-from airbyte.caches.base import SQLCacheInstanceBase, SQLCacheBase
+from airbyte.caches.base import CacheBase
 from airbyte.caches.duckdb import DuckDBCache
-from airbyte._file_writers import JsonlWriterConfig
-from airbyte.caches.base import SQLCacheBase
+from airbyte._processors.file.jsonl import JsonlWriterConfig
+from airbyte.caches.base import CacheBase
 
 
 def test_duck_db_cache_config_initialization():
@@ -28,7 +28,7 @@ def test_get_sql_alchemy_url_with_default_schema_name():
     assert config.get_sql_alchemy_url() == 'duckdb:///test_path'
 
 def test_duck_db_cache_config_inheritance():
-    assert issubclass(DuckDBCache, SQLCacheBase)
+    assert issubclass(DuckDBCache, CacheBase)
     assert issubclass(DuckDBCache, JsonlWriterConfig)
 
 def test_duck_db_cache_config_get_sql_alchemy_url():
@@ -40,7 +40,7 @@ def test_duck_db_cache_config_get_database_name():
     assert config.get_database_name() == 'test_db'
 
 def test_duck_db_cache_base_inheritance():
-    assert issubclass(DuckDBCache, SQLCacheBase)
+    assert issubclass(DuckDBCache, CacheBase)
 
 def test_duck_db_cache_config_default_schema_name():
     config = DuckDBCache(db_path='test_path')
@@ -55,7 +55,7 @@ def test_duck_db_cache_config_get_database_name_with_default_schema_name():
     assert config.get_database_name() == 'test_db'
 
 def test_duck_db_cache_config_inheritance_from_sql_cache_config_base():
-    assert issubclass(DuckDBCache, SQLCacheBase)
+    assert issubclass(DuckDBCache, CacheBase)
 
 def test_duck_db_cache_config_inheritance_from_parquet_writer_config():
     assert issubclass(DuckDBCache, JsonlWriterConfig)
